My progress this week on Ch21, ex 3
I have a model running that seems okay (as best I can tell for now) and have extracted the superpopulation info for the model with all counties (n=85) and n=10 counties. I did not try to do the point estimates for the finite populations but did try to do the 50% interval for the finite populations: hung up on question of how to get equivalent of 'n.sims' in BUGS output from Stan output.