Modified for speed by Eric Goolsby
modified it to implement a faster algorithm -- it computes the log-likelihood using fast tree traversal, so that the you don't ever have to deal with the phylogenetic VCV matrix. Therefore it can handle huge trees very quickly (in fact it's a linear-time algorithm relative to number of species). The method is based on Ho and Ane (Systematic Biology, 2014)